requirements.txt
aio_pika==9.3.0
aiohttp==3.8.5
aiomysql==0.2.0
ccxt==4.0.95
mysql_connector_repackaged==0.3.1
numpy==1.23.5
pandas==2.0.3
PyYAML==6.0.1
PyYAML==6.0.1
rich==13.6.0
